Integrated source, refactored analysis
