+++ TRANSCRIPTION.md
@@ -181,7 +181,7 @@
 ## Words with invalid segments (up to 100 only)
 
 | ID | LANGUAGE | CONCEPT | FORM | SEGMENTS |
-|:---------------------------------|:---------------|:--------------------|:-------------------------|:-------------------------------------------------------------------------------|
+|:---------------------------------|:---------------|:--------------------|:------------------|:-------------------------------------------------------------------------------|
 | CEN-Mundabli-2-115_lizard-2 | CEN-Mundabli-2 | 115_lizard | ŋgɛ̏ː_bɒ́ɦ | ŋ g ɛ ³ <s> <<̏>> </s> <s> <<ː>> </s> b ɒ ⁵ ɦ |
 | CEN-Mundabli-2-11_belly-2 | CEN-Mundabli-2 | 11_belly | kʷű | kʷ <s> <<ű>> </s> |
 | CEN-Mundabli-2-138_pig-2 | CEN-Mundabli-2 | 138_pig | kűːɲam | k <s> <<ű>> </s> <s> <<ː>> </s> ɲ a ³ m |
