+++ cldf/cldf-metadata.json
@@ -12,7 +12,7 @@
         {
             "rdf:about": "https://github.com/kpaam-cam/lowerfungom-wordlists",
             "rdf:type": "prov:Entity",
-            "dc:created": "5c47cd8",
+            "dc:created": "01bcd74",
             "dc:title": "Repository"
         },
         {
