+++ cldf/requirements.txt
@@ -7,19 +7,23 @@ cldfcatalog==1.3.2
 clldutils==3.9.0
 colorlog==5.0.1
 csvw==1.11.0
+cycler==0.10.0
 decorator==4.4.2
 gitdb==4.0.7
 greenlet==1.1.0
 idna==2.10
 iniconfig==1.1.1
 isodate==0.6.0
+kiwisolver==1.3.1
 lingpy==2.6.7
 Markdown==3.3.4
+matplotlib==3.4.2
 networkx==2.5.1
 newick==1.3.0
 numpy==1.20.3
 openpyxl==3.0.7
 packaging==20.9
+pandas==1.2.4
 pluggy==0.13.1
 purl==1.6
 py==1.10.0
@@ -30,10 +34,13 @@ pyconcepticon==2.7.0
 pycountry==20.7.3
 pyglottolog==3.6.0
 pylexibank==3.1.0
+pyparsing==2.4.7
 pytest==6.2.4
+pytz==2021.1
 regex==2021.4.4
 requests==2.25.1
 rfc3986==1.5.0
+scipy==1.7.0
 segments==2.2.0
 six==1.16.0
 smmap==4.0.0
@@ -41,8 +48,10 @@ soupsieve==2.2.1
 SQLAlchemy==1.4.17
 tabulate==0.8.9
 termcolor==1.1.0
+texttable==1.6.3
 tqdm==4.61.0
 uritemplate==3.0.1
 urllib3==1.26.5
+wcwidth==0.2.5
 xlrd==2.0.1
 zenodoclient==0.4.1
