+++ lexibank_kpaamcam.py
@@ -71,5 +71,5 @@ class Dataset(BaseDataset):
                     Value=entry[lid],
                     Language_ID=lid,
                     Parameter_ID=concepts[entry['Concept']],
-                    Source=[]
+                    Source=['Tschonghongei:2021']
                 )
